requirements.txt
asgiref==3.8.1
Django==4.2.26
djangorestframework==3.15.1
mysqlclient==2.2.4
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2